remove unnecessary internal functions and remove loans.rs, util.rs
